refactor: createCouponUseCase 수정 - return -> response dto factory 패턴 적용 - coupon 객체에 expiresAt 저장 및 테스트 코드 수정
